hackernews/pyproject.toml
[tool.poetry]
name = "hackernews"
version = "0.1.0"
description = "parse posts from hackernews"
authors = ["Vitalii Shishorin <moskrc@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1.2"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.5.3"
Flask-RQ = "^0.2"
Flask-SSLify = "^0.1.5"
gunicorn = "^20.0.4"
Flask-Script = "^2.0.6"
flask-marshmallow = "^0.13.0"
flask-restx = "^0.2.0"
beautifulsoup4 = "^4.9.1"
requests = "^2.24.0"
rq-scheduler = "^0.10.0"
flake8 = "^3.8.3"
requests-mock = "^1.8.0"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
yapf = "^0.30.0"
autopep8 = "^1.5.3"
flake8 = "^3.8.3"
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
